Deleted a duplciate folder and fixed a migrations issue
