+++ src/DwitTech.AccountService.WebApi/Controllers/UserController.cs
@@ -2,6 +2,8 @@
 using DwitTech.AccountService.Core.Utilities;
 using Microsoft.AspNetCore.Authorization;
 using DwitTech.AccountService.Core.Dtos;
+using DwitTech.AccountService.Core.Interfaces;
+using Microsoft.AspNetCore.Mvc;
 
 namespace DwitTech.AccountService.WebApi.Controllers
 {
