+++ tests/DwitTech.AccountService.Data.Tests/Repository/UserRepositoryTest.cs
@@ -115,7 +115,6 @@ namespace DwitTech.AccountService.Data.Tests.Repository
                 PhoneNumber = "09023145678",
                 ZipCode = "92001",
                 PostalCode = "Andrew",
-                PassWord = "trionsx",
                 Email = "example@gmail.com",
                 Country = "Brazil",
                 State = "South Casmero",
